[fix]home 필터기능 수정
