+++ Client/src/pages/LandingPage.tsx
@@ -4,9 +4,6 @@ import React, { useState } from 'react';
 
 export default function LandingPage() {
   const [text, setText] = useState('');
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setText(e.target.value);
-  };
   return (
     <>
       <MapContainer />
