+++ Client/src/pages/home/Filter.tsx
@@ -4,7 +4,9 @@ import { RegionInfo } from 'pages/Home';
 import React from 'react';
 import styled from 'styled-components';
 import { Button } from 'component/ui/Button';
-
+import { CRUDdata, ReturnData } from 'api/data';
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 const RegionButton = styled(Button)`
   z-index: 0;
   /* ${ButtonTheme} */
@@ -22,20 +24,34 @@ const RegionButton = styled(Button)`
 
 export default function Filter({
   citys,
-  onClick,
 }: {
   citys: RegionInfo;
-  onClick: (route: string, region?: string) => void;
   region?: string;
+  setState: React.Dispatch<React.SetStateAction<string>>;
+  filter: string;
 }): JSX.Element {
-  console.log(citys);
+  const dd = new CRUDdata();
+  // dd.getData().then(console.log).catch(console.error);
+  const navigate = useNavigate();
 
+  const handleBtnClick = (route: string, region?: string) => {
+    //해당지역으로이동
+    dd.getData()
+      .then((res) => {
+        console.log('tjdrhd');
+        console.log(res);
+
+        navigate(route);
+      })
+      .catch(console.error);
+    navigate(route, { state: {} });
+  };
   return (
     <RegionButton
       children={citys.city}
       className=""
       onClick={() => {
-        onClick(`/board/questionlist`);
+        handleBtnClick(`/board/questionlist`);
       }} //경로랑
       style={{ left: `${citys.x}em`, top: `${citys.y}em` }}
     />
